Add Tensorflow.ipynb file
This file is a prototype system. It loads the data from TrainingSet.
Place this file in TrainingSet folder of Dataset_4NSigComp2010 in order to
run it properly.
Then rescales and binarizes it. It is made following the tutorial
https://www.datacamp.com/community/tutorials/tensorflow-tutorial#gs.Ztfkj8o
TODO:  One hot encode leabels to run it like mnist example
given on tensorflow website.
https://www.tensorflow.org/get_started/mnist/pros